requirements.txt
langchain==0.3.7
langchain-text-splitters==0.3.2
langchain-community==0.3.7
langchain-anthropic==0.3.3
langchain-core==0.3.15
langchain-huggingface==1.2.0
chromadb==0.4.22
sentence-transformers==5.2.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
html2text==2020.1.16
anthropic==0.39.0
python-dotenv==1.0.0
streamlit
